tests, shifted a chunk of test_ajson_ingest to test_utils. it was testing the utils.validate logic essentially.
